+++ bundle/src/main/java/com/adobe/acs/commons/adobeio/service/impl/AdobeioHelper.java
@@ -2,14 +2,14 @@
  * #%L
  * ACS AEM Commons Bundle
  * %%
- * Copyright (C) 2013 - 2018 Adobe
+ * Copyright (C) 2017 Adobe
  * %%
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -17,23 +17,11 @@
  * limitations under the License.
  * #L%
  */
-package com.adobe.acs.commons.adobeio.service;
+package com.adobe.acs.commons.adobeio.service.impl;
 
-import com.adobe.acs.commons.adobeio.types.Action;
-import com.drew.lang.annotations.NotNull;
+import org.apache.http.impl.client.CloseableHttpClient;
 
-import aQute.bnd.annotation.ProviderType;
+public interface AdobeioHelper {
 
-
-/**
- * Interface for the Endpoint service Factory
- */
-@ProviderType
-public interface EndpointServiceFactory {
-
-    /**
-     * @param action Provided action
-     * @return The EndpointService of the provided action
-     */
-    EndpointService getEndpoint(@NotNull Action action);
+    CloseableHttpClient getHttpClient();
 }